requirements.txt
requests~=2.28.1
selenium~=4.8.0
beautifulsoup4==4.11.2
bs4~=0.0.1
python-dotenv~=0.21.1